requirements.txt
certifi==2019.9.11
chardet==3.0.4
idna==2.8
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
six==1.12.0
tzlocal==2.0.0
urllib3==1.25.6
